requirements.txt
langgraph==0.2.14
groq
langchain-google-genai
google-generativeai
requests
beautifulsoup4
lxml
streamlit
fpdf2
python-dotenv
tiktoken
pytest
